Incremental heap census, fixes #7751 (enable with RTS flag -ii)
This patch implements incremental heap census, a way of efficiently
performing a heap census every GC without paying the cost of a major
collection.  The implementation technique is the addition of a
'slopend' field to block descriptors, which divides blocks into
sloppy and non-sloppy portions.  Slop in blocks has already
been taken a census of, and cannot have been GC'd, so it is both
unnecessary and not possible to take a census.

Some back of the envelope measurements:

    (empty)     0.580s
    -hl         0.583s
    -hl -i0     1.739s
    -hl -ii     0.695s

This makes running a heap census every GC an only 20% slowdown,
as opposed to 200% slowdown; an order of magnitude improvement!
(Note that these benchmarks use -hl, so file IO is not being
measured.)

BUGS
    - Breaks 32-bit compilation; we don't have enough fields
      left in bdescr to store slopend!
    - I haven't updated the dump routines to deal with
      incremental census

Signed-off-by: Edward Z. Yang <ezyang@mit.edu>