Generate non-updatable versions of CAFs.
Sometimes we want to make a CAF non-updatable: e.g. if it could generate
infinite work and want to charge it to the user, not the system.
Naively, this requires us to generate *different code* (for example, the
update frame changes the stack layout, do not set the CCS).

The idea is to take every top-level definition that is a CAF, and
duplicate it into a non-updatable version which has the ReEntrant flag
(instead of the Updatable flag).  The original is unchanged, except the
closure now has an extra field (noupd_info) which points to the info
table of the non-updatable version.  User-code can overwrite the info
table of the closure to force it to be non-updatable.

In order to use this properly, you must arrange for the SRT of the
closure you want to overwrite to remain live.  This either means
setting keepCAFs to true, or ensuring that the closure remains live
and not updated (this can be sometimes difficult, as closure evalution
for some built-in CAFs can begin immediately.)

Signed-off-by: Edward Z. Yang <ezyang@mit.edu>