Replace HEAP_ALLOCED with block descriptor check.
Blocks have a new flag BF_STATIC, which indicates if the data
contained therein is static.  This means we can get rid of
all of the complicated caching logic in MBlock.h.  For now,
we still maintain the memory mapping, so we can deallocate
all our memory at the end and do some sanity checks.

Signed-off-by: Edward Z. Yang <ezyang@mit.edu>