Implement compressed static closure code generation.
Summary:
Previously, we wrote data into the binary which resembled the final form
of the static closure on the heap.  However, a lot of the space used
here is unnecessary padding for fields that are only computed at
runtime.  This patch reduces binary sizes by removing those fields,
and only emitting precisely the data we need to reconstruct the object
when we put it on the heap.  This should also be faster, because
of cache effects.

As a further enhancement, we overlap the static closure indirection
table on top of the data (since we no longer need the data after
we copy it to the heap.)

Part of remove HEAP_ALLOCED patch set (#8199)

Depends on D285

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>

Reviewers: simonmar, austin

Subscribers: simonmar, ezyang, carter, thomie

Differential Revision: https://phabricator.haskell.org/D286

GHC Trac Issues: #8199