Properly GC resource containers.
The major change this patch makes is to indirect RCs through a
heap-allocated StgRC object, which is guaranteed to live in the resource
container it refers to.  Thus, as long one of these references is live,
the block associated with the container stays live, so the words held by
the container doesn't hit zero and we don't GC it.

Changes semantics of rcKill: now it's not necessary to kill an RC
in order to make sure it goes away (although if you have outstanding
RCRefs, this is obligatory, because the RCRef will keep the RC alive,
even if the RCRef is not pointing in the container.)

TODO: quantify performance impact of the indirection, maybe
we need another register for StgRC

TODO: prepare_collected_gen should be divided into a single
loop over RC, rather than doing it per gen. Fortunately for minor
GC this does not cost too much.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>