Improve laziness by introducing InitData section type.
Summary:
We now treat static closure indirections and other static
initializer data uniformly: it can be put anywhere in the C--
stream, and it will be gathered up and placed in a single
object file at the end.  You MUST use the InitData section
label; it is no longer correct to simply place your data at
the end of your code generation pipeline.  We still leave
generation of this data in the case of cost centers and HPC
at the end, since it reduces the memory lag between allocation
and usage.

Part of remove HEAP_ALLOCED patch set (#8199)

Depends on D292

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>

Reviewers: simonmar, austin

Subscribers: simonmar, ezyang, carter, thomie

Differential Revision: https://phabricator.haskell.org/D293

GHC Trac Issues: #8199