(WIP) Fix #9422 by maintaining a list of loaded interfaces.
Previously, when we loaded an external interface file, we simply dumped all
of the interfaces it declared into a global environment which was always
consulted.  However, this meant that sometimes, instances could be available
purely by virtue of having been processed earlier than GHC, rather than due
to being transitively reachable by the current module.

This patchset introduces a new piece of state to the TcRnM and IfM monads
which tracks all of the interfaces loaded.  Then, when we do an instance lookup,
we consult this list to see if the interface came from one of these files; if
it didn't, we ignore its existence.

Unfortunately, this patchset doesn't *quite* work yet.  The problem appears
to be, if I type check an interface file, GHC only does this once, but every
time this type checked file is imported, we need to load the knock-on interface
fiels.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>