Handle ghc package's status as wired-in in newer versions of GHC.
When we are building the ghc package from a recent version of GHC
(7.10 or later), we have to make sure the -package-name is what
is expected: it should be 'ghc', not 'ghc-VERSION'.  Note that
we don't make GHC itself wired in for this version of GHC: so when
we build the stage2 compiler its package-name is as before.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>