Generalize packages in installed package database to have multiple units
Summary:
In Backpack, we want to support people writing Cabal packages
like:

    package p where ...
    package q where ...

which a single Cabal package actually defines multiple internal
packages.

This patch gives a name to these subpackages: units, and implements
a small part of the refactoring to give us this.

The main change is in bin-package-db: InstalledPackageInfo has
had many of its GHC-specific fields carved out into InstalledUnitInfo;
and now it contains a list of units.  The subtlety: what goes
where?  My choice: only if a field was *absolutely necessary* to
be per-unit was it placed in a unit; everything else is per-package.
Thus, thinks like extra libraries, ld flags, cc flags, etc. are still
on a per-package basis.  This will simplify later Cabal patches.

We don't actually add any support in ghc-pkg/GHC for handling multiple
units; the minimal API changes to support this were chased out in GHC,
leaning on 'getPrimaryUnit' which assumes that all packages only have
one unit.  This function serves as a marker for pieces of code which
need to be refactored, namely:

    - How we handle searchPackageId for TH package key sanity checks
      and pretty-printing behavior,

    - How we process -package/etc flags which handle packages, not units,

    - How we pretty-print the package database,

    - How we create the package state in the first place.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>

Test Plan: validate

Reviewers: austin, bgamari

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D1057

GHC Trac Issues: #10622