Fix bootstrapping GHC with itself, by distinguishing boot and dist libraries.
Summary:
Previously, we distributed GHC with some base libraries that
had names like "transformers-0.4.3.0".  Then, when we bootstrapped,
we tried to give our bootstrap copies the name "transformers-0.4.3.0".
But this is a conflict and GHC wouldn't have any of it.

The solution is to name the bootstrap copy something different.
We achieve this by having ghc-cabal suffix our component IDs
with dist or dist-boot, depending on what kind of thing it is.

Signed-off-by: Edward Z. Yang <ezyang@cs.stanford.edu>

Test Plan: validate

Reviewers: ggreif, austin, bgamari

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D1355